% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treebar.R
\name{treebar}
\alias{treebar}
\title{Interactive Zoomable Treemap Bar Charts}
\usage{
treebar(data = NULL, ..., width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{\code{string} json data of a four level d3.js hierarchy.
inflexible now but will improve}

\item{...}{additional arguments currently supports 
id, name, and tile for customizing your chart}
}
\description{
htmlwidget based off Chris Given's
\href{http://bl.ocks.org/cmgiven/4541f6de7b6fbef482aaa43f3a71f8d4}{Treemap Bar}
}

